+++ R/sim_data.R
@@ -340,15 +340,16 @@
 #' @param process_data If not `NULL`, it
 #' must be a named list with these
 #' elements: `fun` (required), the function
-#' to further processing the simulated
+#' to further process the simulated
 #' data, such as generating missing data using
 #' functions such as [mice::ampute()]; `args` (optional), a
 #' named list of arguments to be passed
 #' to `fun`, except the one for the
-#' source data; `sim_data_name` (required) the
+#' source data; `sim_data_name` (optional) the
 #' name of the argument to receive the
-#' simulated data (e.g., `data` for
-#' [mice::ampute()]); `processed_data_name`
+#' simulated data (e.g., `"data"` for
+#' [mice::ampute()]), default to
+#' `"data"` if it is not set; `processed_data_name`
 #' (optional), the name of the data frame
 #' after being processed by `fun`,
 #' such as the data frame
@@ -368,6 +369,18 @@
 #' @param ncores The number of CPU
 #' cores to use if parallel processing
 #' is used.
+#'
+#' @param n_ratio If the model is a
+#' multigroup model, and `n` is a single
+#' number, this should be a numeric
+#' vector used to determine the sample
+#' size for each group. For example,
+#' for a two-group model, if `n` is 100
+#' and `n_ratio` is `c(1, 0.5)`, then
+#' the sample sizes for the two groups
+#' are 100 and 50, respectively. If
+#' equal to 1, then all groups have the
+#' same sample size.
 #'
 #' @return
 #' The function [sim_out()] returns
@@ -455,7 +468,8 @@ sim_data <- function(nrep = 10,
                      process_data = NULL,
                      parallel = FALSE,
                      progress = FALSE,
-                     ncores = max(1, parallel::detectCores(logical = FALSE) - 1)) {
+                     ncores = max(1, parallel::detectCores(logical = FALSE) - 1),
+                     n_ratio = 1) {
 
   if (is.null(ptable)) {
     if (is.null(model) || is.null(pop_es)) {
@@ -499,7 +513,8 @@ sim_data <- function(nrep = 10,
                 iseed = iseed,
                 parallel = parallel,
                 progress = progress,
-                ncores = ncores)
+                ncores = ncores,
+                n_ratio = n_ratio)
   class(out) <- c("sim_data", class(out))
   return(out)
 }
@@ -905,7 +920,8 @@ sim_data_i <- function(repid = 1,
                        fit_external = NULL,
                        seed = NULL,
                        drop_list_single_group = TRUE,
-                       merge_groups = TRUE) {
+                       merge_groups = TRUE,
+                       n_ratio = 1) {
   if (!is.null(seed)) set.seed(seed)
   if (is.null(ptable)) {
     ptable <- ptable_pop(model = model,
@@ -924,8 +940,31 @@ sim_data_i <- function(repid = 1,
   }
 
   ngroups <- max(ptable$group)
-  if (length(n) == 1) {
-    n <- rep(n, ngroups)
+  if (ngroups > 1) {
+
+    # ==== Set sample sizes for multigroup models ====
+
+    if (length(n) == 1) {
+
+      # ==== Use n_ratio ====
+
+      if ((length(n_ratio) != ngroups)) {
+        if (length(n_ratio) == 1) {
+          n_ratio <- rep(n_ratio, ngroups)
+        } else {
+          stop("The length of n_ratio is not equal to the number of groups.")
+        }
+      }
+      n <- round(n_ratio * n)
+    } else {
+
+      # ==== n is a vector ====
+
+      if (length(n) != ngroups) {
+        stop("The length of n is not equal to the number of groups.")
+      }
+      # n is used directly
+    }
   }
 
   vnames <- lavaan::lavNames(ptable,
@@ -971,9 +1010,25 @@ sim_data_i <- function(repid = 1,
   model_original <- model
   # add_indicator_syntax() already supports
   # a model syntax with "x:z ~~ y:w"
-  model <- add_indicator_syntax(model,
-                                number_of_indicators = number_of_indicators[[1]],
-                                reliability = reliability[[1]])
+
+  # ==== Check scale scores ====
+
+  tmp <- sapply(mm_lm_dat_out,
+                FUN = colnames,
+                simplify = FALSE)
+  tmp <- unique(unlist(tmp))
+  v_with_indicators <- setdiff(vnames, tmp)
+
+  # ==== Add indicator syntax of for factors replaced by indicators ====
+
+  if (length(v_with_indicators) > 0) {
+    tmp1 <- number_of_indicators[[1]][v_with_indicators]
+    tmp2 <- reliability[[1]][v_with_indicators]
+    model <- add_indicator_syntax(model,
+                                  number_of_indicators = tmp1,
+                                  reliability = tmp2)
+  }
+
   if (!is.null(attr(ptable, "model_fixed"))) {
     if (utils::packageVersion("lavaan") <= "0.6.19") {
       # lavaan 0.6.20+ should support "x:w ~~ y:z"